utils/config.py
"""
Configuration Management

This module handles configuration loading, validation, and management
for the GitHub Repository Analyzer.
"""

import os
import yaml
import json
from typing import Dict, Any, Optional
from pathlib import Path

def load_config(config_path: Optional[str] = None) -> Dict[str, Any]:
    """
    Load configuration from file or environment variables.
    
    Args:
        config_path: Path to configuration file (optional)
        
    Returns:
        Dictionary containing configuration
    """
    config = {}
    
    # Load from file if provided
    if config_path and Path(config_path).exists():
        config.update(load_config_file(config_path))
    
    # Load from environment variables
    config.update(load_env_config())
    
    # Set defaults
    config.update(get_default_config())
    
    return config

def load_config_file(config_path: str) -> Dict[str, Any]:
    """
    Load configuration from file.
    
    Args:
        config_path: Path to configuration file
        
    Returns:
        Dictionary containing configuration
    """
    # TODO: Implement config file loading
    # Support YAML, JSON, and INI formats
    pass

def load_env_config() -> Dict[str, Any]:
    """
    Load configuration from environment variables.
    
    Returns:
        Dictionary containing environment-based configuration
    """
    config = {}
    
    # GitHub API configuration
    config["github"] = {
        "token": os.getenv("GITHUB_TOKEN"),
        "api_base_url": os.getenv("GITHUB_API_BASE_URL", "https://api.github.com")
    }
    
    # LLM configuration
    config["llm"] = {
        "provider": os.getenv("AI_PROVIDER", "openai"),
        "model": os.getenv("AI_MODEL", "gpt-4"),
        "openai_api_key": os.getenv("OPENAI_API_KEY"),
        "anthropic_api_key": os.getenv("ANTHROPIC_API_KEY")
    }
    
    # MCP server configuration
    config["mcp"] = {
        "host": os.getenv("MCP_SERVER_HOST", "localhost"),
        "port": int(os.getenv("MCP_SERVER_PORT", "8000")),
        "timeout": int(os.getenv("MCP_SERVER_TIMEOUT", "30"))
    }
    
    # Streamlit configuration
    config["streamlit"] = {
        "port": int(os.getenv("STREAMLIT_SERVER_PORT", "8501")),
        "address": os.getenv("STREAMLIT_SERVER_ADDRESS", "localhost")
    }
    
    # Rate limiting
    config["rate_limiting"] = {
        "github_rate_limit": int(os.getenv("GITHUB_RATE_LIMIT_PER_HOUR", "5000")),
        "request_timeout": int(os.getenv("REQUEST_TIMEOUT", "30"))
    }
    
    # Logging
    config["logging"] = {
        "level": os.getenv("LOG_LEVEL", "INFO"),
        "file": os.getenv("LOG_FILE", "logs/app.log")
    }
    
    # Cache configuration
    config["cache"] = {
        "enabled": os.getenv("CACHE_ENABLED", "true").lower() == "true",
        "ttl": int(os.getenv("CACHE_TTL", "3600"))
    }
    
    # Repository analysis settings
    config["analysis"] = {
        "max_file_size": int(os.getenv("MAX_FILE_SIZE", "1048576")),
        "max_repository_size": int(os.getenv("MAX_REPOSITORY_SIZE", "100000000")),
        "supported_languages": os.getenv("SUPPORTED_LANGUAGES", "python,javascript,typescript,java,cpp,c,go,rust").split(",")
    }
    
    return config

def get_default_config() -> Dict[str, Any]:
    """
    Get default configuration values.
    
    Returns:
        Dictionary containing default configuration
    """
    return {
        "app": {
            "name": "GitHub Repository Analyzer",
            "version": "1.0.0",
            "debug": False
        },
        "features": {
            "code_analysis": True,
            "visualization": True,
            "smart_summarization": True,
            "change_detection": True,
            "dependency_analysis": True,
            "documentation_generation": True
        }
    }

def validate_config(config: Dict[str, Any]) -> bool:
    """
    Validate configuration values.
    
    Args:
        config: Configuration dictionary to validate
        
    Returns:
        True if configuration is valid, False otherwise
    """
    # TODO: Implement configuration validation
    # Check required fields
    # Validate data types
    # Check for conflicts
    return True

def save_config(config: Dict[str, Any], config_path: str):
    """
    Save configuration to file.
    
    Args:
        config: Configuration dictionary to save
        config_path: Path to save configuration file
    """
    # TODO: Implement configuration saving
    # Support YAML and JSON formats
    pass 