requirements.txt
# 🚀 Core Dependencies
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# 🎨 UI & Styling
streamlit-option-menu>=0.3.6
streamlit-ace>=0.1.1
streamlit-aggrid>=0.3.4
plotly>=5.15.0
altair>=5.0.0

# 🤖 AI & LLM Integration
openai>=1.0.0
anthropic>=0.7.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0

# 🔧 GitHub API & Git
PyGithub>=2.1.0
gitpython>=3.1.0

# 📊 Data Processing & Analysis
pandas>=2.0.0
numpy>=1.24.0
networkx>=3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# 🔍 Code Analysis
ast>=3.8.0
radon>=6.0.0
mccabe>=0.7.0
pylint>=3.0.0
black>=23.0.0

# 🌐 Web & Parsing
beautifulsoup4>=4.12.0
markdown>=3.5.0
pyyaml>=6.0
jinja2>=3.1.0

# 💾 Caching & Storage
redis>=4.5.0
sqlalchemy>=2.0.0

# 🔒 Security & Validation
cryptography>=41.0.0
pydantic[email]>=2.0.0

# 📈 Performance & Monitoring
psutil>=5.9.0
memory-profiler>=0.61.0

# 🧪 Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0

# 🛠️ Development Tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# 📝 Documentation
mkdocs>=1.5.0
mkdocs-material>=9.2.0

# 🌍 Internationalization
babel>=2.12.0

# 📊 Analytics
mixpanel>=4.10.0

# 🔄 Async & Concurrency
aiohttp>=3.8.0
asyncio-mqtt>=0.16.0

# 🎯 MCP Protocol (when available)
# mcp>=1.0.0

# 📱 Mobile & PWA Support
streamlit-pwa>=0.1.0 